pyproject.toml
[tool.poetry]
name = "ai-ops-agents"
version = "0.1.0"
description = ""
authors = ["Karim Bisri <bisri.karim@gmail.com>"]
packages = [{ include = "ai_ops_agents", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
langchain = "0.2.0"
chromadb = "^1.3.4"
nats-py = "^2.12.0"
kubernetes = "^34.1.0"
pydantic = "^2.12.4"
gitpython = "^3.1.45"
python-dotenv = "^1.2.1"
fastapi = "^0.121.2"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
typer = "^0.20.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
